+++ .env.sample
@@ -1,20 +1,20 @@
 # .env.sample
 
 # Telegram Bot Token для интеграции с Telegram
-TELEGRAM_BOT_TOKEN=your_telegram_bot_token
+TELEGRAM_BOT_TOKEN=your_telegram_bot_token  # Замените на реальный токен
 
-# Данные для подключения к базе данных (если используешь PostgreSQL)
-DB_NAME=your_db_name
-DB_USER=your_db_user
-DB_PASSWORD=your_db_password
-DB_HOST=localhost
+# Данные для подключения к базе данных (если используете PostgreSQL)
+DB_NAME=your_db_name  # Замените на название вашей базы данных
+DB_USER=your_db_user  # Замените на имя пользователя базы данных
+DB_PASSWORD=your_db_password  # Замените на ваш пароль
+DB_HOST=localhost  # Замените на имя хоста контейнера или IP-адрес
 DB_PORT=5432
 
 # Настройки для Celery и Redis
-CELERY_BROKER_URL=redis://localhost:6379/0
-CELERY_RESULT_BACKEND=redis://localhost:6379/0
+CELERY_BROKER_URL=redis://localhost:6379/0  # Замените на правильный хост Redis
+CELERY_RESULT_BACKEND=redis://localhost:6379/0  # Замените на правильный хост Redis
 
 # Прочие переменные окружения
-SECRET_KEY=your_secret_key
+SECRET_KEY=your_secret_key  # Замените на реальный секретный ключ
 DEBUG=True
-ALLOWED_HOSTS=127.0.0.1, .localhost
+ALLOWED_HOSTS=127.0.0.1,localhost  # Добавьте ваши доменные имена или IP для продакшн
