+++ config/views.py
@@ -1,3 +1,6 @@
-from django.shortcuts import render
+# tracker/views.py
+from django.http import HttpResponse
 
-# Create your views here.
+
+def home(request):
+    return HttpResponse("<h1>Welcome to the Habit Tracker API</h1>")
