--- tracker/urls.py
@@ -1,47 +1,0 @@
-"""
-URL configuration for tracker project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
-from django.contrib import admin
-from django.urls import path, include
-from django.views.generic import RedirectView  # Добавляем RedirectView
-from drf_yasg.views import get_schema_view
-from drf_yasg import openapi
-from rest_framework import permissions
-
-# Создаем схему для документации
-schema_view = get_schema_view(
-   openapi.Info(
-      title="Habits API",
-      default_version='v1',
-      description="Документация для API приложения привычек",
-      terms_of_service="https://www.google.com/policies/terms/",
-      contact=openapi.Contact(email="contact@habits.local"),
-      license=openapi.License(name="BSD License"),
-   ),
-   public=True,
-   permission_classes=[permissions.AllowAny],  # Используем список для разрешений
-)
-
-urlpatterns = [
-    path("admin/", admin.site.urls),
-    path("api/", include("habits.urls")),  # Подключаем API для привычек
-    path("api/users/", include("users.urls")),  # Подключаем API для пользователей
-    path("", RedirectView.as_view(url="/api/")),  # Редирект на API
-    # Эндпоинты для генерации Swagger и ReDoc документации
-    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
-    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
-]
