--- tracker/views.py
@@ -1,6 +1,0 @@
-# tracker/views.py
-from django.http import HttpResponse
-
-
-def home(request):
-    return HttpResponse("<h1>Welcome to the Habit Tracker API</h1>")
